requirements.txt
anyio==4.2.0
asgiref==3.7.2
certifi==2023.11.17
crispy-bootstrap5==2023.10
deprecation==2.1.0
Django==5.0
django-ckeditor==6.7.0
django-crispy-forms==2.1
django-email-verification==0.3.3
django-js-asset==2.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
packaging==23.2
Pillow==10.1.0
PyJWT==2.8.0
python-telegram-bot==20.7
sniffio==1.3.0
sqlparse==0.4.4
validators==0.22.0
